refactor | dtos

+++ domain/src/main/java/com/saludsystem/submodules/movimiento/model/dtos/command/VentaDetalleCreateCommand.java
@@ -3,6 +3,7 @@ package com.saludsystem.submodules.movimiento.model.dtos.command;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @AllArgsConstructor
@@ -21,9 +22,9 @@ public class VentaDetalleCreateCommand {
     private Integer cantidad;
 
     @Schema(example = "0")
-    private Double precioUnitario;
+    private BigDecimal precioUnitario;
 
     @Schema(example = "0")
-    private Double subtotal;
+    private BigDecimal subtotal;
 
 }
